Added GPU versions of multiple objective functions.
- reg:linear, reg:logistic, binary:logistic and binary:logitraw have GPU versions now
