+++ src/objective/objective.cc
@@ -28,6 +28,9 @@ namespace xgboost {
 namespace obj {
 // List of files that will be force linked in static links.
 DMLC_REGISTRY_LINK_TAG(regression_obj);
+#ifdef XGBOOST_USE_CUDA
+  DMLC_REGISTRY_LINK_TAG(regression_obj_gpu);
+#endif
 DMLC_REGISTRY_LINK_TAG(multiclass_obj);
 DMLC_REGISTRY_LINK_TAG(rank_obj);
 }  // namespace obj
